Starting a story based on what you hear on the radio
It has been a NPR program, and for some reason they were asking a question whose answer was "rub a baby with lard". It so happens that doctors actually did it, sometime in the past or even now. So I mixed my fascination with hobo culture and stuff and came up with this title and beginning